flutter/packages/simple_brilliant_app/examples/vision_api/server/requirements.txt
fastapi
uvicorn
pillow